requirements.txt
flask>=2.3.5
streamlit>=1.25.0
openai-whisper>=20230314
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.26.0
pandas>=2.1.0
spacy>=3.7.0
sentence-transformers>=2.2.2
joblib>=1.3.2
scikit-learn>=1.3.0
pydub>=0.25.1
soundfile>=0.12.1
ffmpeg-python>=0.2.1